scripts/requirements.txt
azure
beautifulsoup4
lxml
Pillow
praw
python-dateutil
requests
six
update-checker